moved Collaborate folder
from RadicalCollaboration to pages
