+++ src/pages/Collaborate/index.js
@@ -3,9 +3,9 @@ import React from 'react';
 import Box from '@material-ui/core/Box';
 import Grid from '@material-ui/core/Grid';
 import Container from '@material-ui/core/Container';
-import Link from '../../../components/common/Link';
-import BottomCallToAction from '../../../components/BottomCallToAction';
-import { GenericHeaderSection } from '../../../components/';
+import Link from '../../components/common/Link';
+import BottomCallToAction from '../../components/BottomCallToAction';
+import { GenericHeaderSection } from '../../components/';
 import makeStyles from '@material-ui/core/styles/makeStyles';
 import LeftTextRightImage from './LeftTextRightImage';
 import RightTextLeftImage from './RightTextLeftImage';
